requirements.txt
twitchio==2.7.0
fastapi==0.110.0
uvicorn==0.29.0
python-dotenv==1.0.1
pydantic==2.7.1
